Opportunistically accept unknown value types
When morphing values from source to wireformat, which is uploaded to
Kubernetes, this plugin performs type coherence checking to ensure
you're not uploading "foo" into a field that expects a number.

Previously, if the value you were trying to pass was "Unknown", meaning
it was to be computed during the apply phase, you would get an error
about failing to morph from object element to object element. This is a
good error to have in place since it prevents the user from accidentally
passing an inappropriately typed value into the Kubernetes cluster and
doesn't even require communication with the cluster, making it decidable
very early in the execution flow.

This change modifies the algorithm to allow situations where you're not
dependent on type coercion to proceed. E.g. the "Unknown" value has a
known type that is compatible with the destination type and therefore it
doesn't matter what the value is for purposes of type compatibility.

It is still allowed to pass a string into a number if that value is
numeric, but only if that value is known at plan time.

Fixed: #190