Add support for non-structured custom resources (#203)
* Add support for non-structured custom resources

Non-structured custom resources have CRDs which have an
empty schema. This means their body outside of the metadata
can be any arbitrary structure and we can't infer it from
the OpenAPI spec for the cluster. In this case we make a
best effort attempt to guess the type information from the
config.

* Do a dry-run for non-structured resources