Use tftypes ValidateValue to catch panics (#239)
Use ValidateValue prior to calling NewValue, as suggested by the docs:

```
If the passed value is not a valid value for the passed type, NewValue will panic. 
Any value and type combination that does not return an error from ValidateValue 
is guaranteed to not panic. When calling NewValue with user input with a type not
 known at compile time, it is recommended to call ValidateValue before calling
 NewValue, to allow graceful handling of the error.
```